Rearrange QM driver files
