+++ src/qibolab/instruments/qm/__init__.py
@@ -1,0 +1,3 @@
+from qibolab.instruments.qm.config import QMPort
+from qibolab.instruments.qm.driver import QMOPX
+from qibolab.instruments.qm.simulator import QMSim
